+++ src/index.ts
@@ -1,6 +1,6 @@
-import { Layer, Psd, ReadOptions, WriteOptions } from './psd';
+import { Psd, ReadOptions, WriteOptions } from './psd';
 import { PsdWriter, writePsd as writePsdInternal, getWriterBuffer, createWriter, getWriterBufferNoCopy } from './psdWriter';
-import { PsdReader, readPsd as readPsdInternal, createReader, decodeLayerImageData } from './psdReader';
+import { PsdReader, readPsd as readPsdInternal, createReader, decodeLayerPixels, getLayerImageData, getLayerMaskImageData, getLayerRealMaskImageData, getLayerCanvas, getLayerMaskCanvas, getLayerRealMaskCanvas, getCompositeImageData, getCompositeCanvas } from './psdReader';
 import { fromByteArray } from 'base64-js';
 
 export * from './abr';
@@ -16,6 +16,8 @@ interface BufferLike {
 }
 
 export const byteArrayToBase64 = fromByteArray;
+
+export { decodeLayerPixels, getLayerImageData, getLayerMaskImageData, getLayerRealMaskImageData, getLayerCanvas, getLayerMaskCanvas, getLayerRealMaskCanvas, getCompositeImageData, getCompositeCanvas };
 
 export function readPsd(buffer: ArrayBuffer | BufferLike, options?: ReadOptions): Psd {
 	const reader = 'buffer' in buffer ?
@@ -42,8 +44,4 @@ export function writePsdBuffer(psd: Psd, options?: WriteOptions): Buffer {
 	}
 
 	return Buffer.from(writePsdUint8Array(psd, options));
-}
-
-export function decodeLayerPixels(layer: Layer, useImageData?: boolean) {
-	decodeLayerImageData(layer, useImageData);
 }